+++ SENT_backend/internal/models/mongo_models.go
@@ -13,7 +13,7 @@ import (
 type SoftwareItem struct {
 	ID              primitive.ObjectID `bson:"_id,omitempty" json:"id"`
 	AssetHWID       string             `bson:"asset_hwid" json:"asset_hwid" binding:"required,max=64"`
-	OrgID           int64              `bson:"org_id" json:"org_id" binding:"required"`
+	OrgID           uint               `bson:"org_id" json:"org_id" binding:"required"`
 	SoftwareName    string             `bson:"software_name" json:"software_name" binding:"required,min=1,max=255"`
 	Version         string             `bson:"version" json:"version" binding:"max=50"`
 	Publisher       string             `bson:"publisher" json:"publisher" binding:"max=255"`
@@ -26,7 +26,7 @@ type SoftwareItem struct {
 type OpenPort struct {
 	ID          primitive.ObjectID `bson:"_id,omitempty" json:"id"`
 	AssetHWID   string             `bson:"asset_hwid" json:"asset_hwid" binding:"required,max=64"`
-	OrgID       int64              `bson:"org_id" json:"org_id" binding:"required"`
+	OrgID       uint               `bson:"org_id" json:"org_id" binding:"required"`
 	Port        int                `bson:"port" json:"port" binding:"required,min=1,max=65535"`
 	ProcessName string             `bson:"process_name" json:"process_name" binding:"max=255"`
 	Status      string             `bson:"status" json:"status" binding:"required,oneof=OPEN CLOSED"`
@@ -36,7 +36,7 @@ type OpenPort struct {
 type USBLog struct {
 	ID            primitive.ObjectID `bson:"_id,omitempty" json:"id"`
 	AssetHWID     string             `bson:"asset_hwid" json:"asset_hwid" binding:"required,max=64"`
-	OrgID         int64              `bson:"org_id" json:"org_id" binding:"required"`
+	OrgID         uint               `bson:"org_id" json:"org_id" binding:"required"`
 	DeviceName    string             `bson:"device_name" json:"device_name" binding:"required,min=1,max=255"`
 	DeviceID      string             `bson:"device_id" json:"device_id" binding:"max=255"`
 	VID           string             `bson:"vid" json:"vid" binding:"max=10"`
@@ -51,7 +51,7 @@ type USBLog struct {
 type AssetIOActivity struct {
 	ID               primitive.ObjectID `bson:"_id,omitempty" json:"id"`
 	AssetHWID        string             `bson:"asset_hwid" json:"asset_hwid" binding:"required,max=64"`
-	OrgID            int64              `bson:"org_id" json:"org_id" binding:"required"`
+	OrgID            uint               `bson:"org_id" json:"org_id" binding:"required"`
 	NetBytesSent     uint64             `bson:"net_bytes_sent" json:"net_bytes_sent" binding:"required,gte=0"`
 	NetBytesRecv     uint64             `bson:"net_bytes_recv" json:"net_bytes_recv" binding:"required,gte=0"`
 	DiskBytesWritten uint64             `bson:"disk_bytes_written" json:"disk_bytes_written" binding:"required,gte=0"`
@@ -62,7 +62,7 @@ type AssetIOActivity struct {
 type AssetInventory struct {
 	ID         primitive.ObjectID `bson:"_id,omitempty" json:"id"`
 	AssetHWID  string             `bson:"asset_hwid" json:"asset_hwid" binding:"required,max=64"`
-	OrgID      int64              `bson:"org_id" json:"org_id" binding:"required"`
+	OrgID      uint               `bson:"org_id" json:"org_id" binding:"required"`
 	CPUModel   string             `bson:"cpu_model" json:"cpu_model" binding:"required,min=1,max=255"`
 	RAMTotalGB int                `bson:"ram_total_gb" json:"ram_total_gb" binding:"required,min=0,max=1000000"`
 	OSInfo     string             `bson:"os_info" json:"os_info" binding:"required,min=1,max=500"`
@@ -70,9 +70,9 @@ type AssetInventory struct {
 }
 type SecurityAlert struct {
 	ID          primitive.ObjectID `bson:"_id,omitempty" json:"id"`
-	OrgID       int64              `bson:"org_id" json:"org_id" binding:"required"`
+	OrgID       uint               `bson:"org_id" json:"org_id" binding:"required"`
 	AssetHWID   string             `bson:"asset_hwid" json:"asset_hwid" binding:"required,max=64"`
-	IncidentID  *int64             `bson:"incident_id,omitempty" json:"incident_id"`
+	IncidentID  *uint              `bson:"incident_id,omitempty" json:"incident_id"`
 	Priority    string             `bson:"priority" json:"priority" binding:"required,oneof=P1 P2 P3 P4"`
 	AlertType   string             `bson:"alert_type" json:"alert_type" binding:"required,max=50"`
 	Title       string             `bson:"title" json:"title" binding:"required,min=1,max=255"`
@@ -85,8 +85,8 @@ type SecurityAlert struct {
 type IncidentAudit struct {
 	ID          primitive.ObjectID `bson:"_id,omitempty" json:"id"`
 	CreatedAt   time.Time          `bson:"created_at" json:"created_at"`
-	IncidentID  int64              `bson:"incident_id" json:"incident_id" binding:"required"`
-	UserID      *int64             `bson:"user_id,omitempty" json:"user_id"`
+	IncidentID  uint               `bson:"incident_id" json:"incident_id" binding:"required"`
+	UserID      *uint              `bson:"user_id,omitempty" json:"user_id"`
 	User        *User              `bson:"-" json:"user,omitempty"`
 	UserName    string             `bson:"user_name" json:"user_name"`
 	ActionType  string             `bson:"action_type" json:"action_type" binding:"required,max=50"`
@@ -104,7 +104,7 @@ type IncidentAudit struct {
 }
 
 func (act *IncidentAudit) GenerateAuditHash() {
-	var uid int64 = 0
+	var uid uint = 0
 	if act.UserID != nil {
 		uid = *act.UserID
 	}
